+++ Calculator/Calculator/Model/ExpressionParser.swift
@@ -30,4 +30,3 @@ enum ExpressionParser {
         return input.split(with: " ")
     }
 }
-
