--- Calculator/Calculator/Tests/CalculatorTests.swift
@@ -1,20 +1,0 @@
-//
-//  CalculatorTests.swift
-//  CalculatorTests
-//
-//  Created by 김동욱 on 2022/03/14.
-//
-
-import XCTest
-@testable import Calculator
-
-class CalculatorTests: XCTestCase {
-
-    override func setUpWithError() throws {
-        try super.setUpWithError()
-    }
-
-    override func tearDownWithError() throws {
-        try super.tearDownWithError()
-    }
-}
